Keep code DRY by merging transaction components
